library: Stop relying on LUT_MASK
It was only used in sinLerp and cosLerp because in the very early days
of the library libnds didn't have the functions and the arrays SIN_LUT
and COS_LUT were used directly. When libnds was updated and it stopped
exposing the arrays Nitro Engine had to adapt. The masking wasn't needed
back then either, it was a mistake, but I forgot to remove it then.

In the case of the noise screen effect, it was just being misused.