library: Introduce safe dual 3D mode
Regular dual 3D mode is very unstable. As soon as you drop one frame,
the last screen to be drawn will be displayed on both screens.

This new system is a lot more complex to setup, but it's reliable, and
it doesn't need a third main VRAM bank for triple buffering, just VRAM I
to act as pseudo framebuffer.

The old dual 3D mode has been kept for compatibility.

Documentation will come soon.

Thanks to @Gericom for the tips!