library: Swap active screen in dual 3D mode correctly
At the moment, in dual 3D mode, the VBL handler swaps the active screen.
This is wrong, because if the game framerate drops below 60 FPS the
active screen will go out of sync with the main loop (assuming that this
is where NE_ProcessDual() is called.

It makes more sense to modify the flag whenever NE_ProcessDual() is
called.

Note that dropping below 60 FPS may still desync the screens, even
swapping the top and bottom screens. This patch doesn't fix that issue,
it just opens the door to a future fix.