library: Fix dual 3D DMA mode in BlocksDS
The fork of libnds of BlocksDS doesn't leave 32 unused characters right
at the beginning of the tileset of the debug console. The code used to
rely on this space being free to store there the first line of a 16-bit
bitmap.

The line has been moved to an actual free part of the VRAM bank.