library: Use safe DMA functions in BlocksDS
They aren't available in the libnds provided by devkitPro. I could copy
the safe versions of the functions to Nitro Engine, but that wouldn't
help as the program would still use the unsafe versions for everything
else outside of Nitro Engine. The fix needs to happen in libnds for it
to work.