img2ds: Deprecate tool for regular textures
It's unlikely this tool will be removed from the repository, but it's no
longer going to be maintained now that the version of grit included in
BlocksDS has had a few bugs related to NDS textures fixed.

The main purpose for img2ds was to eventually support compressed
textures, but ptexconv already does a good job at it.