library: Refactor sprite setup in dual 3D modes
Instead of allocating the buffer permanently, only allocate it while the
sprites are setup. Then, free the buffer.

Also, the cache needs to be flushed before doing a DMA copy. This used to
cause black squares on the screen because of some sprite attributes not
being properly copied to OAM.